refactor: move files in core to correct dirs
