+++ lib/core/domain/ambee_api/repository/pollen_repository.dart
@@ -1,4 +1,4 @@
-import 'package:pollen_meter/core/domain/models/pollen_model.dart';
+import '../models/pollen_model.dart';
 
 abstract class PollenRepository {
   Future<PollenModel> fetchData({
